Refactor: Split OptionsPopup into multiple classes
# Conflicts:
#	core/src/com/unciv/ui/options/OptionsPopup.kt
